requirements.txt
alembic==1.13.2
bidict==0.23.1
blinker==1.8.2
blosc2==2.7.1
certifi==2024.7.4
cffi==1.17.0
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
dnspython==2.6.1
email_validator==2.2.0
eventlet==0.36.1
Flask==3.0.3
Flask-Migrate==4.0.7
Flask-SocketIO==5.3.6
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.2.1
gevent==24.2.1
greenlet==3.0.3
h11==0.14.0
idna==3.7
itsdangerous==2.2.0
Jinja2==3.1.4
Mako==1.3.5
MarkupSafe==2.1.5
msgpack==1.0.8
ndindex==1.8
numexpr==2.10.1
numpy==1.26.4
packaging==24.1
pandas==2.2.2
py-cpuinfo==9.0.0
pycparser==2.22
python-dateutil==2.9.0.post0
python-engineio==4.9.1
python-socketio==5.11.3
pytz==2024.1
requests==2.32.3
scipy==1.14.0
setuptools==72.2.0
simple-websocket==1.0.0
six==1.16.0
SQLAlchemy==2.0.32
tables==3.10.0
typing_extensions==4.12.2
tzdata==2024.1
urllib3==2.2.2
Werkzeug==3.0.3
wsproto==1.2.0
WTForms==3.1.2
zope.event==5.0
zope.interface==7.0.1
